Moved get_ring_properties from lattice_object to utils
